pyproject.toml
[project]
name = "social-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "django-silk>=5.4.3",
    "django-unfold>=0.71.0",
    "djlint>=1.36.4",
    "easy-thumbnails>=2.10.1",
    "pillow>=12.0.0",
    "pyopenssl>=25.3.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "social-auth-app-django>=5.6.0",
    "werkzeug>=3.1.3",
]
